Fix OOM when having a lot of recurrence rules
We had a broken client that created calendar entries with a lot of
recurrence rule (all of them FREQ=WEEKLY) and a lot of alarms. This led
to SOGo going out of memory.

This patch works around that by filtering out duplicated rules as they
yield the same date ranges and we only need to have them once.